+++ src/test/java/org/cicirello/search/operators/reals/GaussianMutationTests.java
@@ -45,66 +45,6 @@ public class GaussianMutationTests {
 	private static final int MAX_TRIALS = 100;
 	
 	@Test
-	public void testEquals() {
-		GaussianMutation<RealValued> g1 = GaussianMutation.createGaussianMutation(1);
-		GaussianMutation<RealValued> g2 = GaussianMutation.createGaussianMutation(1);
-		GaussianMutation<RealValued> g3 = GaussianMutation.createGaussianMutation(2);
-		assertEquals(g1, g2);
-		assertEquals(g1.hashCode(), g2.hashCode());
-		assertNotEquals(g1, g3);
-		assertFalse(g1.equals(null));
-		assertFalse(g1.equals("hello"));
-		GaussianMutation<RealValued> p1 = GaussianMutation.createGaussianMutation(1, 1);
-		GaussianMutation<RealValued> p2 = GaussianMutation.createGaussianMutation(2, 1);
-		GaussianMutation<RealValued> p3 = GaussianMutation.createGaussianMutation(1, 2);
-		GaussianMutation<RealValued> p4 = GaussianMutation.createGaussianMutation(1, 0.5);
-		GaussianMutation<RealValued> p5 = GaussianMutation.createGaussianMutation(1, 0.25);
-		assertNotEquals(p1, p2);
-		assertNotEquals(p1, p3);
-		assertNotEquals(p4, p5);
-		assertNotEquals(p1, g1);
-		assertNotEquals(p1, g3);
-		assertNotEquals(p2, p1);
-		assertNotEquals(p3, p1);
-		assertNotEquals(p5, p4);
-		assertNotEquals(g1, p1);
-		assertNotEquals(g3, p1);
-		assertFalse(p1.equals(null));
-		p1 = UndoableGaussianMutation.createGaussianMutation(1, 1);
-		p2 = UndoableGaussianMutation.createGaussianMutation(2, 1);
-		p3 = UndoableGaussianMutation.createGaussianMutation(1, 2);
-		p4 = UndoableGaussianMutation.createGaussianMutation(1, 0.5);
-		p5 = UndoableGaussianMutation.createGaussianMutation(1, 0.25);
-		assertNotEquals(p1, p2);
-		assertNotEquals(p1, p3);
-		assertNotEquals(p4, p5);
-		assertNotEquals(p4, p1);
-		assertNotEquals(p1, g1);
-		assertNotEquals(p1, g3);
-		assertNotEquals(p2, p1);
-		assertNotEquals(p3, p1);
-		assertNotEquals(p5, p4);
-		assertNotEquals(g1, p1);
-		assertNotEquals(g3, p1);
-		assertFalse(p1.equals(null));
-		GaussianMutation<RealValued> b1 = GaussianMutation.createGaussianMutation(1.0, 1.0, 5.0);
-		GaussianMutation<RealValued> b2 = GaussianMutation.createGaussianMutation(1.01, 1.0, 5.0);
-		GaussianMutation<RealValued> b3 = GaussianMutation.createGaussianMutation(1.0, 1.01, 5.0);
-		GaussianMutation<RealValued> b4 = GaussianMutation.createGaussianMutation(1.0, 1.0, 5.01);
-		GaussianMutation<RealValued> b1Copy = GaussianMutation.createGaussianMutation(1.0, 1.0, 5.0);
-		assertEquals(b1, b1Copy);
-		assertEquals(b1.hashCode(), b1Copy.hashCode());
-		assertNotEquals(b1, b2);
-		assertNotEquals(b1, b3);
-		assertNotEquals(b1, b4);
-		assertNotEquals(b2, b1);
-		assertNotEquals(b3, b1);
-		assertNotEquals(b4, b1);
-		assertNotEquals(b1, p1);
-		assertFalse(b1.equals(null));
-	}
-	
-	@Test
 	public void testToArray() {
 		GaussianMutation<RealValued> u = GaussianMutation.createGaussianMutation(2);
 		double[] a = u.toArray(null);
@@ -171,30 +111,26 @@ public class GaussianMutationTests {
 		
 		UndoableGaussianMutation<RealValued> g5copyU = g5u.split();
 		assertEquals(5.0, g5copyU.get(0), EPSILON);
-		assertEquals(g5u, g5copyU);
-		assertEquals(g5u.hashCode(), g5copyU.hashCode());
+		assertEquals(g5u.get(0), g5copyU.get(0));
 		verifyMutate1(g5copyU);
 		verifyUndo(g5copyU);
 		
 		g5copyU = g5u.copy();
 		assertEquals(5.0, g5copyU.get(0), EPSILON);
-		assertEquals(g5u, g5copyU);
-		assertEquals(g5u.hashCode(), g5copyU.hashCode());
+		assertEquals(g5u.get(0), g5copyU.get(0));
 		verifyMutate1(g5copyU);
 		verifyUndo(g5copyU);
 		
 		GaussianMutation<RealValued> g5split = g5.split();
 		assertEquals(5.0, g5split.get(0), EPSILON);
-		assertEquals(g5, g5split);
+		assertEquals(g5.get(0), g5split.get(0));
 		assertTrue(g5 != g5split);
-		assertEquals(g5.hashCode(), g5split.hashCode());
 		verifyMutate1(g5split);
 		
 		GaussianMutation<RealValued> g5copyM = g5.copy();
 		assertEquals(5.0, g5copyM.get(0), EPSILON);
-		assertEquals(g5, g5copyM);
+		assertEquals(g5.get(0), g5copyM.get(0));
 		assertTrue(g5 != g5copyM);
-		assertEquals(g5.hashCode(), g5copyM.hashCode());
 		verifyMutate1(g5copyM);
 		
 		GaussianMutation<RealValued> g3 = g5;
@@ -344,30 +280,26 @@ public class GaussianMutationTests {
 			
 			UndoableGaussianMutation<RealValued> g5copy = g5.copy();
 			assertEquals(5.0, g5copy.get(0), EPSILON);
-			assertEquals(g5, g5copy);
-			assertEquals(g5.hashCode(), g5copy.hashCode());
+			assertEquals(g5.get(0), g5copy.get(0));
 			verifyMutate1(g5copy, k);
 			verifyUndo(g5copy);
 			
 			GaussianMutation<RealValued> g5copyDis = g5dis.copy();
 			assertEquals(5.0, g5copyDis.get(0), EPSILON);
-			assertEquals(g5dis, g5copyDis);
-			assertEquals(g5dis.hashCode(), g5copyDis.hashCode());
+			assertEquals(g5dis.get(0), g5copyDis.get(0));
 			verifyMutate1(g5copyDis, k);
 			
 			UndoableGaussianMutation<RealValued> g5split = g5.split();
 			assertEquals(5.0, g5split.get(0), EPSILON);
-			assertEquals(g5, g5split);
+			assertEquals(g5.get(0), g5split.get(0));
 			assertTrue(g5 != g5split);
-			assertEquals(g5.hashCode(), g5split.hashCode());
 			verifyMutate1(g5split, k);
 			verifyUndo(g5split);
 			
 			GaussianMutation<RealValued> g5splitDis = g5dis.split();
 			assertEquals(5.0, g5splitDis.get(0), EPSILON);
-			assertEquals(g5dis, g5splitDis);
+			assertEquals(g5dis.get(0), g5splitDis.get(0));
 			assertTrue(g5dis != g5splitDis);
-			assertEquals(g5dis.hashCode(), g5splitDis.hashCode());
 			verifyMutate1(g5splitDis, k);
 			
 			UndoableGaussianMutation<RealValued> g3 = g5;
@@ -394,30 +326,26 @@ public class GaussianMutationTests {
 			
 			UndoableGaussianMutation<RealValued> g5copy = g5.copy();
 			assertEquals(5.0, g5copy.get(0), EPSILON);
-			assertEquals(g5, g5copy);
-			assertEquals(g5.hashCode(), g5copy.hashCode());
+			assertEquals(g5.get(0), g5copy.get(0));
 			verifyMutate1(g5copy, k);
 			verifyUndo(g5copy);
 			
 			GaussianMutation<RealValued> g5copyDis = g5dis.copy();
 			assertEquals(5.0, g5copyDis.get(0), EPSILON);
-			assertEquals(g5dis, g5copyDis);
-			assertEquals(g5dis.hashCode(), g5copyDis.hashCode());
+			assertEquals(g5dis.get(0), g5copyDis.get(0));
 			verifyMutate1(g5copyDis, k);
 			
 			UndoableGaussianMutation<RealValued> g5split = g5.split();
 			assertEquals(5.0, g5split.get(0), EPSILON);
-			assertEquals(g5, g5split);
+			assertEquals(g5.get(0), g5split.get(0));
 			assertTrue(g5 != g5split);
-			assertEquals(g5.hashCode(), g5split.hashCode());
 			verifyMutate1(g5split, k);
 			verifyUndo(g5split);
 			
 			GaussianMutation<RealValued> g5splitDis = g5dis.split();
 			assertEquals(5.0, g5splitDis.get(0), EPSILON);
-			assertEquals(g5dis, g5splitDis);
+			assertEquals(g5dis.get(0), g5splitDis.get(0));
 			assertTrue(g5dis != g5splitDis);
-			assertEquals(g5dis.hashCode(), g5splitDis.hashCode());
 			verifyMutate1(g5splitDis, k);
 			
 			UndoableGaussianMutation<RealValued> g3 = g5;
@@ -431,7 +359,7 @@ public class GaussianMutationTests {
 	
 	
 	
-	private void verifyMutate1(GaussianMutation<RealValued> m) {
+	private void verifyMutate1(AbstractRealMutation<RealValued> m) {
 		int countLow = 0;
 		int countHigh = 0;
 		for (int i = 0; i < MAX_TRIALS && (countLow==0 || countHigh==0); i++) {
@@ -467,7 +395,7 @@ public class GaussianMutationTests {
 		}
 	}
 	
-	private void verifyMutate1(GaussianMutation<RealValued> m, double p) {
+	private void verifyMutate1(AbstractRealMutation<RealValued> m, double p) {
 		int countLow = 0;
 		int countHigh = 0;
 		final int TRIALS = (int)(2 * MAX_TRIALS / p);
@@ -506,7 +434,7 @@ public class GaussianMutationTests {
 		}
 	}
 	
-	private void verifyMutate1(GaussianMutation<RealValued> m, int K) {
+	private void verifyMutate1(AbstractRealMutation<RealValued> m, int K) {
 		int countLow = 0;
 		int countHigh = 0;
 		for (int i = 0; i < MAX_TRIALS && (countLow==0 || countHigh==0); i++) {