+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-09-23
+## [Unreleased] - 2022-09-26
 
 ### Added
 * AdaptiveEvolutionaryAlgorithm class: An implementation of an evolutionary algorithm with control 
@@ -17,7 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * Refactored evolutionary algorithm classes to improve maintainability, as well as a few minor optimizations.
-* Refactored Gaussian mutation.
+* Refactored mutation classes for real-valued representations, including:
+  * Extracted an internal abstract base class for common functionality.
+  * Refactored Gaussian mutation classes to use new abstract base class.
 
 ### Deprecated